Update @typechain/ethers-v5 to have strict contract types
The update to `@typechain/ethers-v5` includes
https://github.com/ethereum-ts/TypeChain/pull/373. This ensures that the
interfaces generated for contracts are typed and we can’t call
arbitrary functions on contracts.

The update also requires us to update `typechain` package from `v4` to
`v5`. This results in changes to `hardhat.config.ts`.

Finally, we add dependencies that are required by the generated
contract bindings.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>